Se agregaron mas tests, se modifico la clase EjecutadorDelPrograma para que presente la infomación de una mejor manera y se organizaron algunos detalles faltantes como nombres e informción del README

+++ src/main/java/org/santiago_ricardo/service/EstudianteRecordService.java
@@ -3,17 +3,17 @@ package org.santiago_ricardo.service;
 public interface EstudianteRecordService {
     double calcularPromedio();
     int contarEstudiantes();
-    int contarEstudiantesQueNoDeseanEstudiar();
-    int contarEstudiantesSinRecursos();
-    int contarEstudiantesQueRecibenEsenanzasAntiguas();
-    int contarEstudiantesSinRecursosYNoDeseanEstudiar();
-    int contarEstudiantesQueNoDeseanEstudiarYRecibenEsenanzasAntiguas();
-    int contarEstudiantesSinRecursosYRecibenEsenanzasAntiguas();
-    int contarEstudiantesSinRecursosYRecibenEsenanzasAntiguasYNoDeseanEstudiar();
-    int calcularEdadQueMasSeRepite();
+    double contarEstudiantesQueNoDeseanEstudiar();
+    double contarEstudiantesSinRecursos();
+    double contarEstudiantesQueRecibenEsenanzasAntiguas();
+    double contarEstudiantesSinRecursosYNoDeseanEstudiar();
+    double contarEstudiantesQueNoDeseanEstudiarYRecibenEsenanzasAntiguas();
+    double contarEstudiantesSinRecursosYRecibenEsenanzasAntiguas();
+    double contarEstudiantesSinRecursosYRecibenEsenanzasAntiguasYNoDeseanEstudiar();
+    double calcularEdadQueMasSeRepite();
     //Promedio bajo se refiere a un promedio inferior a 3
-    int contarPromedioBajoYRecibenEsenanzasAntiguas();
-    int contarPromedioBajoYNoDeseanEstudiar();
+    double contarPromedioBajoYRecibenEsenanzasAntiguas();
+    double contarPromedioBajoYNoDeseanEstudiar();
     String[] encontrarMejores3PromediosDepartamentos();
     String[] encontrarPeores3PromediosDepartamentos();
     String encontrarDepartamentoConMayorCantidadEstudiantesNoDeseanEstudiar();
